requirements-gcloud.txt
# Google Cloud 優化版依賴清單 (排除problematic packages)

# 基礎框架
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic[email]==2.8.2
pydantic-settings==2.3.4

# 數據獲取
yfinance==0.2.28
pandas==2.2.2
numpy==1.26.4
requests==2.32.3

# 視覺化 (排除scipy和scikit-learn以避免編譯錯誤)
matplotlib==3.9.1
plotly==5.17.0
kaleido==0.2.1

# 數據庫
sqlalchemy==2.0.31

# 認證與安全
PyJWT==2.8.0
python-multipart==0.0.9
bcrypt==4.1.3
passlib==1.7.4
python-jose[cryptography]==3.3.0

# OAuth
google-auth==2.29.0
google-auth-oauthlib==1.2.0

# AI集成
openai==1.35.13

# 基礎工具
python-dotenv==1.0.1
aiofiles==23.2.1
python-dateutil==2.9.0.post0
pytz==2024.1

# 異步支持
aiohttp==3.9.5
websockets==12.0
httpx==0.27.0

# 日誌與監控
loguru==0.7.2

# 圖像處理
Pillow==10.4.0